move wishlist to cog and add addition/removal via react
