+++ main.py
@@ -22,7 +22,6 @@ from commands.speak import speak, speak_embed
 from commands.trekduel import trekduel
 from commands.trektalk import trektalk
 from commands.tuvix import tuvix
-from commands.wishlist import *
 
 # Slash Command Groups
 import commands.drop
@@ -49,6 +48,7 @@ from cogs.slots import Slots
 from cogs.trade import Trade
 from cogs.react_roles import ReactRoles
 from cogs.backups import Backups
+from cogs.wishlist import Wishlist
 from cogs.wordcloud import Wordcloud
 bot.add_cog(ChaosZork(bot))
 bot.add_cog(Poker(bot))
@@ -59,6 +59,7 @@ bot.add_cog(Shop(bot))
 bot.add_cog(Slots(bot))
 bot.add_cog(Trade(bot))
 bot.add_cog(Backups(bot))
+bot.add_cog(Wishlist(bot))
 bot.add_cog(Wordcloud(bot))
 if config["roles"]["reaction_roles_enabled"]:
   bot.add_cog(ReactRoles(bot))
